Document getting started, sort out components and sidebar
