+++ .storybook/config.js
@@ -1,4 +1,20 @@
-import { configure } from '@storybook/react';
+import { addParameters, configure } from '@storybook/react';
 
-// automatically import all files ending in *.stories.js
-configure(require.context('../stories', true, /\.stories\.tsx$/), module);
+const sectionOrder = {
+  'rehowl': 0,
+  'hooks': 1,
+  'components': 2
+}
+
+const getStorySection =
+  story => sectionOrder[story[1].id.split('-')[0]]
+
+addParameters({
+  options: {
+    hierarchySeparator: /\./,
+    storySort: (a, b) =>
+      a[1].kind === b[1].kind ? 0 : getStorySection(a) - getStorySection(b),
+  }
+})
+
+configure(require.context('../stories', true, /\.stories\.(tsx|mdx)$/), module);
