+++ stories/useHowl.stories.mdx
@@ -1,0 +1,36 @@
+import { Meta, Props, Description, Story, Preview } from '@storybook/addon-docs/blocks'
+import { useState } from 'react'
+import { useHowl } from '../src'
+import sound1 from './static/audio/sound1.mp3'
+
+
+<Meta title="Hooks|useHowl" component={useHowl} />
+
+# useHowl
+
+<Description of={useHowl} />
+
+```js
+function MyComponent() {
+  const {
+    howl,
+    state,
+    error,
+    load
+  } = useHowl({
+    src: './mysound.mp3'
+  })
+  // howl can be handed to <Play />
+}
+```
+
+Values provided by useHowl are:
+
+| Variable | Description |
+|------|------|
+| howl | Howl instance |
+| state | 'unloaded', 'loading', or 'loaded' |
+| error | { message, id } of load error, if any. |
+| load | Call this function to load source if you set `preload` to false. |
+
+<Props of={useHowl} />
